Hokey rename step because of windows
